Add tile_def_info() helper
This helper works like tile_gui_info(), tile_player_info() et al,
except that it takes a tile_def structure, calling whichever
texture-specific function is appropriate to retrieve the tile texture info.

It's a little strange that this is a method of the ImageManager; ideally
the ImageManager would query this function rather than maintain a map of
texture to texture-info-getters (which is how this function is currently
implemented).