Add UI widget event observer API
This commit adds a way to register callbacks for UI-related events.
This mechanism is intended to facilitate coordination between widgets
(as opposed to sub-classing, which customizes behavior within a given
widget).

Listeners are type-safe, and can be anything convertible to a
std::function. Most widgets won't have any listeners attached, so
callbacks are stored in structs associated with the widget type,
rather than within the target widget itself.