Use menu for yesno prompt outside game (Currency)
This fixes an issue where the yesno prompt text was not being shown
because redraw_screen() was not being called to show updates to the
message buffer. Rather than work out what's actually going on there,
I've opted to use the TOUCH_UI popup code when outside of a game, as
showing the messages panel without a game loaded looks very buggy.

mprf() is still used when calling yesno() inside a game, unless TOUCH_UI
is defined.